Update patch set 1
Patch Set 1:

nice, but just for my understanding: when are we supposed to use the existing attribute scheme, e.g. voip_usr_preferences etc. and when can we introduce separate tables like you did? is it because of the unique constraints you need or the grouped settings like email+regex tuples?

Patch-set: 1