Update patch set 1
Patch Set 1:

a general question: as the local timezone is most times not set explicitly for output/return vals, how did you find out its wrong with this case of the sms journal? actually it should show the same times as others (eg. some create timestamp happened at the same time as an sms). if not, is it maybe because of the time was inserted with a differing time?

Patch-set: 1