TT#17458 Remove unnecessary or obsolete transitive dependencies
Several of the dependencies were just transitional ones, required by
libsipwise-base-perl, but that module now properly depends on anything
it requires itself, so we do not need to specify them here. In addition
some were remnants from when that module was using those dependencies,
but that is not the case anymore.

Change-Id: I40f701bad5a66dad0e6cf3d1984472f557a093f0