Update patch set 4
Patch Set 4:

>Oh and in some cases «git log -p», even though some of these seem to have been present already before the git history started.

ok was curious if you know a autoamted way. i did not checked in detail if one of those removed ones could cause an issues, but for this repo i think it's fine.
there are however others codes that load dynamically or "require" stuff somewhere inside code. examples i remember are cpuaffinity in bulk-processor or the maketext /merge mechanisms in panel scripts. so particularly with panel it can be dangerous.

also another thing worht to mention: i would not change the particular depndecies for the json libs -> there are bugs the with the module load mechanism and boolean constants (xs vs pp), we had pain to make it work as is (also tests).

Patch-set: 4